tests/wpt/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/004.html

<!doctype html>
<title>Reuse of document object after document.open</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="/common/blank.html"></iframe>
<script>
var t = async_test();
var iframe;
onload = t.step_func(function() {
  var iframe = document.getElementsByTagName("iframe")[0];
  var handle = iframe.contentDocument;
  iframe.contentDocument.test_state = 1;
  assert_equals(iframe.contentDocument.open(), handle);
  assert_equals(iframe.contentDocument.test_state, 1);
  assert_equals(iframe.contentDocument, handle);
  t.done();
});
</script>
